batch_prediction_pipeline/pyproject.toml
[tool.poetry]
name = "batch-prediction-pipeline"
version = "0.1.0"
description = ""
authors = ["Massyl GH <massyl-6@hotmail.fr>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
hopsworks = "^4.1.5"
google-cloud-storage = "^2.19.0"
fire = "^0.7.0"
joblib = "^1.4.2"
python-dotenv = "^1.0.1"
pyarrow = "^18.1.0"
sktime = "^0.16.1"
training_pipeline = "0.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "test"
url = "http://localhost:8080/"  # URL used to download your packages from